### 2019-12-04 #1934 @dmey
@danielskatz I think this may be due to the 2 affiliations (see https://github.com/teb-model/teb/blame/joss-paper/paper/paper.md#L15).
