### 2019-12-04 #1934 @danielskatz
👋 @dmey - the multiple affiliations line needs to be of the form
affiliation: "1, 2"
with quote marks, not square brackets

Once you've made this change, you can enter a new comment in this thread with ```@whedon generate pdf from branch joss-paper``` to try again.  Feel free to iterate until things look right.
