+++ paper/paper.bib
@@ -308,8 +308,7 @@
 }
 
 @software{meyer_d_2019_3562311,
-  author       = {Meyer, D. and
-                  Raustad, R.},
+  author       = {Meyer, D. and Raustad, R.},
   title        = {MinimalDX},
   month        = dec,
   year         = 2019,
