+++ paper/paper.bib
@@ -314,7 +314,7 @@
   year         = 2019,
   publisher    = {Zenodo},
   version      = {0.1.0},
-  doi          = {10.5281/zenodo.3562310,
+  doi          = {10.5281/zenodo.3562310},
   url          = {https://doi.org/10.5281/zenodo.3562310}
 }
 
