+++ paper/paper.bib
@@ -13,7 +13,7 @@
   pages = {357-397}
 }
 
-@article{lemonsu_2004_J.Appl.Meteorol.,
+@article{lemonsu_2004_J.Appl.Meteorol,
   title = {Modeling the {{Surface Energy Balance}} of the {{Core}} of an {{Old Mediterranean City}}: {{Marseille}}},
   volume = {43},
   language = {en},
@@ -23,7 +23,7 @@
   pages = {16}
 }
 
-@article{lemonsu_2012_Geosci.ModelDev.,
+@article{lemonsu_2012_Geosci.ModelDev,
   title = {Inclusion of Vegetation in the {{Town Energy Balance}} Model for Modelling Urban Green Areas},
   volume = {5},
   issn = {1991-9603},
@@ -37,7 +37,7 @@
   pages = {1377-1393}
 }
 
-@article{lac_2018_Geosci.ModelDev.,
+@article{lac_2018_Geosci.ModelDev,
   title = {Overview of the {{Meso}}-{{NH}} Model Version 5.4 and Its Applications},
   volume = {11},
   issn = {1991-9603},
@@ -51,7 +51,7 @@
   pages = {1929-1969}
 }
 
-@article{demunck_2013_Int.J.Climatol.,
+@article{demunck_2013_Int.J.Climatol,
   title = {How Much Can Air Conditioning Increase Air Temperatures for a City like {{Paris}}, {{France}}?},
   volume = {33},
   issn = {08998418},
@@ -65,7 +65,7 @@
   pages = {210-227}
 }
 
-@article{bueno_2012_Geosci.ModelDev.,
+@article{bueno_2012_Geosci.ModelDev,
   title = {Development and Evaluation of a Building Energy Model Integrated in the {{TEB}} Scheme},
   volume = {5},
   issn = {1991-9603},
@@ -79,7 +79,7 @@
   pages = {433-448}
 }
 
-@article{schoetter_2017_Geosci.ModelDev.,
+@article{schoetter_2017_Geosci.ModelDev,
   title = {Parametrisation of the Variety of Human Behaviour Related to Building Energy Consumption in the {{Town Energy Balance}} ({{SURFEX}}-{{TEB}} v. 8.2)},
   volume = {10},
   issn = {1991-9603},
@@ -108,7 +108,7 @@
   pages = {323-344}
 }
 
-@article{masson_2014_Front.Environ.Sci.,
+@article{masson_2014_Front.Environ.Sci,
   title = {Solar Panels Reduce Both Global Warming and Urban Heat Island},
   volume = {2},
   issn = {2296-665X},
@@ -187,7 +187,7 @@
   pages = {1468}
 }
 
-@article{rozoff_2003_J.Appl.Meteorol.,
+@article{rozoff_2003_J.Appl.Meteorol,
   title = {Simulation of {{St}}. {{Louis}}, {{Missouri}}, {{Land Use Impacts}} on {{Thunderstorms}}},
   volume = {42},
   language = {en},
@@ -197,7 +197,7 @@
   pages = {23}
 }
 
-@article{pielke_1992_Meteorl.Atmos.Phys.,
+@article{pielke_1992_Meteorl.Atmos.Phys,
   title = {A Comprehensive Meteorological Modeling System \textemdash{} {{RAMS}}},
   volume = {49},
   issn = {0177-7971, 1436-5065},
@@ -211,15 +211,25 @@
   pages = {69-91}
 }
 
-@article{xue_2000_,
[…]
