### 2019-12-04 #1934 @dmey
@danielskatz I have fixed the references but there is still an issue with the typesetting of footnotes. It looks like table footnotes are made part of the table itself and not of the page footer.  Have you experienced a similar issue before?
