+++ paper/paper.bib
@@ -191,7 +191,7 @@
   title = {Simulation of {{St}}. {{Louis}}, {{Missouri}}, {{Land Use Impacts}} on {{Thunderstorms}}},
   volume = {42},
   language = {en},
-  journal = {JOURNAL OF APPLIED METEOROLOGY},
+  journal = {Journal of Applied Meteorology},
   author = {Rozoff, Christopher M and Cotton, William R and Adegoke, Jimmy O},
   year = {2003},
   pages = {23}
@@ -340,7 +340,7 @@ url="https://doi.org/10.1007/s007030070003"
   title = {Evaluation of the {{Town Energy Balance}} ({{TEB}}) {{Scheme}} with {{Direct Measurements}} from {{Dry Districts}} in {{Two Cities}}},
   volume = {41},
   language = {en},
-  journal = {JOURNAL OF APPLIED METEOROLOGY},
+  journal = {Journal of Applied Meteorology},
   author = {Masson, Val{\'e}ry and Grimmond, C S B and Oke, T R},
   year = {2002},
   pages = {16}
