### 2019-12-04 #1934 @arfon
> @arfon - Any thoughts?

No idea sorry. I think this is probably pushing the limit of the Markdown/Pandoc template sorry. I'd recommend trying to find a different way to accomplish this.
