### 2019-12-04 21041e0
remove footnote in table [skip ci]
