### 2019-12-04 #1934 @dmey
> No idea sorry. I think this is probably pushing the limit of the Markdown/Pandoc template sorry. I'd recommend trying to find a different way to accomplish this.

@arfon  @danielskatz  Thanks and not problems -- I removed the footnote and used a start (*) instead this with clarification in the caption. LGTM now but please let me know for any issues.
