### 2019-12-05 #1934 @dmey
@mjsottile and @danielskatz many thanks for this -- In the meantime, I would like to suggest a few names [from the JOSS list of reviewers](https://bit.ly/joss-reviewers) that may be suitable for the review:

```
https://github.com/dvalters
https://github.com/bjarketol
https://github.com/tomchor
https://github.com/kramea
https://github.com/martinfleis
https://github.com/robertodr
https://github.com/nmstreethran
https://github.com/johnrobertlawson
https://github.com/brynpickering
```
