### 2019-12-15 #1934 @tomchor
@mjsottile Unfortunately I don't have enough expertise or time to review this paper at this moment. So I'll have to decline.

Thank you.
