### 2020-01-08 #1934 @mjsottile
:wave: @zbeekman @ethan-nelson @mdelorme @edoddridge Would any of you be interested in reviewing this submission for JOSS?
