### 2020-01-09 #2008 @whedon
[REVIEW]: Enhanced software and platform for the Town Energy Balance (TEB) model
**Submitting author:** @dmey (<a href="http://orcid.org/0000-0002-7071-7547">D. Meyer</a>)
**Repository:** <a href="https://github.com/teb-model/teb" target ="_blank">https://github.com/teb-model/teb</a>
**Version:** 4.0.0
