### 2020-01-09 40dd6d1
add missing DOIs
