+++ paper/paper.bib
@@ -18,6 +18,7 @@
   volume = {43},
   language = {en},
   journal = {Journal of Applied Meteorology},
+  doi = {10.1175/1520-0450(2004)043<0312:MTSEBO>2.0.CO;2},
   author = {Lemonsu, A and Grimmond, C S B and Masson, V},
   year = {2004},
   pages = {16}
@@ -284,6 +285,7 @@ url="https://doi.org/10.1007/s007030070003"
   title = {The {{Meso}}-{{NH Atmospheric Simulation System}}. {{Part I}}: Adiabatic Formulation and Control Simulations},
   language = {en},
   journal = {Annales Geophysicae},
+  doi = {10.1007/s00585-997-0090-6},
   author = {Lafore, J P and Stein, J and Asencio, N and Bougeault, P and Ducrocq, V and Duron, J and Fischer, C and Masson, V and Pinty, J P and Redelsperger, J L and Richard, E},
   year = {1998},
   pages = {20}
@@ -341,8 +343,8 @@ url="https://doi.org/10.1007/s007030070003"
   volume = {41},
   language = {en},
   journal = {Journal of Applied Meteorology},
+  doi = {10.1175/1520-0450(2002)041<1011%3AEOTTEB>2.0.CO%3B2},
   author = {Masson, Val{\'e}ry and Grimmond, C S B and Oke, T R},
   year = {2002},
   pages = {16}
-
 }
