### 2020-01-09 #2008 @dmey
@mjsottile, @rasmussn, @ethan-nelson I have added the three missing DOIs. Will keep the summary paper on a separate branch (`joss-paper`) in case of any changes and will only merge it into `master` once you are all happy with it.
