### 2020-01-24 #2008 @ethan-nelson
This is my first review with JOSS, so thanks for this opportunity. Code looks to be in good shape. Tests run fine and physical values in the raw output of testing are reasonable. Modules are documented on the corresponding website, model scientific documentation is available, and namelist parameters are defined. The paper covers the important points of the software and highlights where it aligns with existing packages.

Responding to a few “unchecked” points in the checklist that are minor:

## General checks

* License: The OSI list only indicates that CECILL-2.1 is open source compatible. It looks like there are some differences between CECILL-1, and I'm not sure if the OSI designation is backwards-compatible, so I will defer to @mjsottile for that.

## Functionality

* Installation: (CentOS 7.7, cmake 3.12.2, gcc 4.8.5) I ran into issues during compilation:

```
/home/enelson/review/teb/external/minimal-dx/src/shared/minimal_dx_fan.f90:99.64:

      error stop 'Fan Mode must either be 1 for on or 0 for off'
                                                                1
Error: ERROR STOP statement not allowed in PURE procedure at (1)
gmake[2]: *** [CMakeFiles/teb.dir/external/minimal-dx/src/shared/minimal_dx_fan.f90.o] Error 1
gmake[1]: *** [CMakeFiles/teb.dir/all] Error 2
gmake: *** [all] Error 2
```

After removing the error message, the model compiled just fine. This is a fairly old gcc version though so it may not warrant a change.

## Documentation

* Automated tests: this is a corner case, but it is useful if the user could specify a custom cmake/gcc/git binary location for tests, either in the python call or near the top of one of the python scripts. My cmake is not in my PATH variable so I had to manually change the location in the python code. Adding this would allow for faster iteration over different versions of gcc, etc. for testing compatibility.

* Community guidelines: Might I suggest a CONTRIBUTING.md file or indication in the README to complement the journal text? That could help direct folks who stumble upon the repo first as opposed to the paper.

## Other feedback:

* A file in `docs/` with brief documentation detailing the necessary `input/` forcing files would be useful for tinkering from scratch. I ended up copying files from the example to play around with, but it would be good to have a list all the file variables and units--similar to the namelist docs.

* Something else to consider: what are your thoughts on writing out a simple metadata file in the `output/` directory of a given run? Perhaps this onus should fall on the user to document their runs, but I could see potentially pulling in some metadata from the model build time (e.g. `git rev-parse`) that you could copy to `output/META.txt` or the like.
