### 2020-02-19 #2008 @irnaka
Sorry, this is not my area of expertise.

==============================
Theodosius Marwan IRNAKA

*marwan.irnaka@gmail.com <marwan.irnaka@gmail.com>*
*+62852 2803 5489 <%2B6285228035489>*
+336 47 08 81 33
* <marwan.irnaka@mail.ugm.ac.id>*

On Wed, 19 Feb 2020 at 22:19, Matthew Sottile <notifications@github.com>
wrote:

> 👋 @RCHG <https://github.com/RCHG> @irnaka <https://github.com/irnaka>
> @jayten <https://github.com/jayten> Hello - we are looking for an
> additional reviewer for this JOSS submission to replace a reviewer who can
> no longer do a review. I found your names on the potential reviewers
> spreadsheet based on your preferred languages and domains of expertise.
> Would any of you be available for a review?
>
