### 2020-02-21 #2008 @jayten
I have gone through the review process for the [`teb`](https://github.com/teb-model/teb) repository.  I was able to install (on Ubuntu18 and MacOs 10.13) and run test cases with instructions provided in the documentation. The software solves the intended problem well. The software is well written and documented. However, the flavor of the FORTRAN code style is old, owing to the use of old legacy code. For example, "impnone.fh" just contain an 'implicit none' statement. Although code style is not a major issue, it can be improved over time with the contribution of others. I would suggest @dmey to add, at least a simple,  **CONTRIBUTING.md** for interested individuals to contribute to [`teb`](https://github.com/teb-model/teb).

# License
I have issues with the license file: [CeCILL-C-V1](https://github.com/teb-model/teb/blob/joss-paper/Licence_CeCILL-C_V1-en.txt). The license file mentioned is not a part of the [OSI](https://opensource.org/licenses/alphabetical). Additionally, some old source files (like '[modei_flxsurf3db.f](https://github.com/teb-model/teb/blob/joss-paper/src/teb/modi_flxsurf3bx.f)') mention different license in the header, which is also not part of the [OSI](https://opensource.org/licenses/alphabetical). I want @mjsottile to take a decision on this issue.

# Paper
@dmey , make sure all the contributors are at least acknowledged in the paper.md. For example, **P. Le Moigne**, mentioned as the author of the [*abort1_sfx.f90*](https://github.com/teb-model/teb/blob/joss-paper/src/driver/abor1_sfx.F90) source file, is neither mentioned in the author list nor the acknowledgment.
## References
Missing DOI for bib entry: rozoff_2003_J.Appl.Meteorol
