### 2020-02-21 #2008 @mjsottile
@jayten Thanks for the great and prompt review.  I will look into the issues you raise when I return from travel next week.  I need to look more into the CeCILL licenses - as I understand them from looking at them a number of years ago, they are generally recognized as similar to the FSF and OSI licenses, but originated in Europe independently of the more common licenses that largely originated in the US.  I’ll do some reading and possibly ask the other JOSS editors for their opinions on the issue.

In the meantime @dmey, if you would like to look at the issues related to contributors and reference DOIs, that would be useful.
