### 2020-02-21 #2008 @danielskatz
JOSS requires submitted software to have an OSI-approved license.  It looks like we (actually I) missed this originally, for which I apologize to the author and reviewers.
