### 2020-02-21 #2008 @danielskatz
👋 @dmey - Is it possible to change (or dual license) with an OSI-approved license?  If not, we will have to ask you to withdraw this submission or we will have to reject it.
