### 2020-02-21 #2008 @danielskatz
Note that https://opensource.org/licenses/CECILL-2.1 is an OSI-approved license - perhaps CeCILL-C-V1 is as well, but I can't find it on the OSI site.
