### 2020-03-02 #2008 @dmey
@mjsottile and @danielskatz I have discussed these points with Masson given that he is the co-author who developed the physical model in 2000 and wrote most of the original code. We should be able to make most of the changes as requested but would to clarify a couple of points with you first before going ahead with the changes:

1.	Licence: OK -- this can be upgraded to CECILL-2.1 but will ask Masson to approve this as it will affect the whole software and not only my changes/enhancements.
2.	Acknowledgments: in the repository there are a couple of files authored by Le Moigne. I have double checked this again with Masson but it looks like these files were included from a 'sister' model (SURFEX) and Le Moigne was not a TEB contributor. Do you still require this to be acknowledged in the summary paper?
3.	Zenodo: We had not yet deposited the software on Zenodo however, given that my contribution is recent and specifically about the enhancements as described in the paper, I think that the software version reference should reflect all the software contributors including those acknowledged in the summary paper and with a different order – i.e. Masson et al. (2020).
