### 2020-03-02 #2008 @danielskatz
The following are my thoughts as the AEiC on duty when this came up.  @mjsottile and the currently on-duty @openjournals/jose-eics might also have thoughts.

> 1. Licence: OK -- this can be upgraded to CECILL-2.1 but will ask Masson to approve this as it will affect the whole software and not only my changes/enhancements.

A license change needs to be approved by all contributors in theory, since they are the copyright holders.

> 2. Acknowledgments: in the repository there are a couple of files authored by Le Moigne. I have double checked this again with Masson but it looks like these files were included from a 'sister' model (SURFEX) and Le Moigne was not a TEB contributor. Do you still require this to be acknowledged in the summary paper?

If this software was not developed for the project but is included (in some sense, as a dependency), it likely makes sense to acknowledge that dependency. The author of this dependent software would not need to be an author of this software, in my opinion.

> 3. Zenodo: We had not yet deposited the software on Zenodo however, given that my contribution is recent and specifically about the enhancements as described in the paper, I think that the software version reference should reflect all the software contributors including those acknowledged in the summary paper and with a different order – i.e. Masson et al. (2020).

I'm not sure I understand this.  Both the JOSS paper and associated Zenodo (or other) archival repository deposit should contain the same list of authors, in the other determined by the authors, and with the list being determined by the authors.  The authors should include all significant software contributors, plus potentially others who contributed to the project in ways that may not be captured in the repository (e.g. software design, user support, ...)
