### 2020-03-02 #2008 @dmey
> I'm not sure I understand this. Both the JOSS paper and associated Zenodo (or other) archival repository deposit should contain the same list of authors, in the other determined by the authors, and with the list being determined by the authors. The authors should include all significant software contributors, plus potentially others who contributed to the project in ways that may not be captured in the repository (e.g. software design, user support, ...)

I appreciate that this may be a bit confusing -- the reason I have asked is because we would rather not fork the original model but instead, improve and move the model to GitHub (i.e. this paper) and continue future development from this point forward. Therefore, whereas in the paper the authorship is clear for the work that was done in improving TEB, I think that for the associated Zenodo archive, it would be proper to keep the software name and authorship to reflect the contributions made over the years of development (i.e. leave name as TEB -- as also reflected in GitHub repository-- thus avoiding the creation of a separate fork for this specific submission). This is something that has been discussed with the TEB developers and seemed to make most sense.
