### 2020-03-03 #2008 @arfon
If I understand the situation, the paper will have a shorter author list than the Zenodo archive? I don't think this is ideal but I think it's probably OK to allow this given the justification.
