### 2020-03-04 #2008 @mjsottile
I agree.  It seems that the only outstanding item is to make sure all necessary parties consent to the update from CeCILL1 to CeCILL-2.1 such that the license will conform to the JOSS requirements.  Once you've indicated that is complete, I can scan through the state of the submission to push it to the next stage if it looks ready.
