### 2020-03-05 #2008 @mjsottile
@dmey In addition to the license issue, both reviewers have not checked off the box related to community guidelines.  It looks like the paper provides guidance for users wishing to contribute, but the README or other documents in the repository do not.  You should add some text to the README or a separate CONTRIBUTING.md file for users who encounter the package but do not read the JOSS paper.

Also, it looks like there is an invalid DOI in the paper - can you repair that?  (See the "check references" result above).

Once you've had a chance to address these issues, please take a look at the reviewer comments again to make sure you've addressed both of their concerns.  At that point I'll check in with them again to see if they're satisfied with the outstanding boxes from their checklists.
