### 2020-03-11 c44de60
begin updating licence headers to CeCILL 2.1
