### 2020-03-11 681883c
update SFX_LIC to CeCILL version 2.1 as per email
