### 2020-03-11 fa3969e
Update software licence to CeCILL version 2.1 (#23)
Updates the licence to CeCILL version 2.1 in header for source files and updates the two licence files at the root Licence_CeCILL-C_V1-en.txt to Licence_CeCILL_V2.1-en.txt and Licence_CeCILL-C_V1-fr.txt to Licence_CeCILL_V2.1-fr.txt.
