### 2020-03-11 44d5710
Travis: fix macOS
