+++ .travis.yml
@@ -27,19 +27,18 @@ matrix:
         apt:
           packages:
             - gfortran
-    - language: generic
+    - language: minimal
       os: osx
       osx_image: xcode11.2
       env: BUILD_TYPE=Debug
-    - language: generic
+    - language: minimal
       os: osx
       osx_image: xcode11.2
       env: BUILD_TYPE=Release
 
 before_install:
-  - if [[ $TRAVIS_OS_NAME == osx ]]; then brew update && brew upgrade gcc python; fi
-  - which python3
-  - python --version
+  - gfortran --version
+  - python3 --version
   - python3 -m pip install -r requirements.txt
 
 script:
