+++ paper/paper.bib
@@ -339,12 +339,12 @@ url="https://doi.org/10.1007/s007030070003"
 }
 
 @article{masson_2002_J.Appl.Meteorol,
-  title = {Evaluation of the {{Town Energy Balance}} ({{TEB}}) {{Scheme}} with {{Direct Measurements}} from {{Dry Districts}} in {{Two Cities}}},
-  volume = {41},
-  language = {en},
-  journal = {Journal of Applied Meteorology},
-  doi = {10.1175/1520-0450(2002)041<1011%3AEOTTEB>2.0.CO%3B2},
-  author = {Masson, Val{\'e}ry and Grimmond, C S B and Oke, T R},
-  year = {2002},
-  pages = {16}
+author = {Masson, V. and Grimmond, C. S. B. and Oke, T. R.},
+title = {Evaluation of the Town Energy Balance (TEB) Scheme with Direct Measurements from Dry Districts in Two Cities},
+journal = {Journal of Applied Meteorology},
+volume = {41},
+number = {10},
+pages = {1011-1026},
+year = {2002},
+doi = {10.1175/1520-0450(2002)041<1011:EOTTEB>2.0.CO;2}
 }
