+++ paper/paper.bib
@@ -1,3 +1,43 @@
+@article{Redon2020,
+  doi = {10.5194/gmd-13-385-2020},
+  url = {https://doi.org/10.5194/gmd-13-385-2020},
+  year = {2020},
+  month = feb,
+  publisher = {Copernicus {GmbH}},
+  volume = {13},
+  number = {2},
+  pages = {385--399},
+  author = {Emilie Redon and Aude Lemonsu and Val{\'{e}}ry Masson},
+  title = {An urban trees parameterization for modeling microclimatic variables and thermal comfort conditions at street level with the Town Energy Balance model ({TEB}-{SURFEX} v8.0)},
+  journal = {Geoscientific Model Development}
+}
+
+@article{Redon2017,
+  doi = {10.5194/gmd-10-385-2017},
+  url = {https://doi.org/10.5194/gmd-10-385-2017},
+  year = {2017},
+  month = jan,
+  publisher = {Copernicus {GmbH}},
+  volume = {10},
+  number = {1},
+  pages = {385--411},
+  author = {Emilie C. Redon and Aude Lemonsu and Val{\'{e}}ry Masson and Benjamin Morille and Marjorie Musy},
+  title = {Implementation of street trees within the solar radiative exchange parameterization of {TEB} in {SURFEX} v8.0},
+  journal = {Geoscientific Model Development}
+}
+
+@article{Pigeon2014,
+  doi = {10.1016/j.enbuild.2013.10.038},
+  url = {https://doi.org/10.1016/j.enbuild.2013.10.038},
+  year = {2014},
+  month = jun,
+  publisher = {Elsevier {BV}},
+  volume = {76},
+  pages = {1--14},
+  author = {G. Pigeon and K. Zibouche and B. Bueno and J. Le Bras and V. Masson},
+  title = {Improving the capabilities of the Town Energy Balance model with up-to-date building energy simulation algorithms: an application to a set of representative buildings in Paris},
+  journal = {Energy and Buildings}
+}
 
 @article{masson_2000_Boundary-LayerMeteorology,
   title = {A {{Physically}}-{{Based Scheme For The Urban Energy Budget In Atmospheric Models}}},
