+++ paper/paper.bib
@@ -229,13 +229,14 @@
 }
 
 @article{rozoff_2003_J.Appl.Meteorol,
-  title = {Simulation of {{St}}. {{Louis}}, {{Missouri}}, {{Land Use Impacts}} on {{Thunderstorms}}},
-  volume = {42},
-  language = {en},
+  author = {Rozoff, Christopher M. and Cotton, William R. and Adegoke, Jimmy O.},
+  title = {Simulation of St. Louis, Missouri, Land Use Impacts on Thunderstorms},
   journal = {Journal of Applied Meteorology},
-  author = {Rozoff, Christopher M and Cotton, William R and Adegoke, Jimmy O},
+  volume = {42},
+  number = {6},
+  pages = {716-738},
   year = {2003},
-  pages = {23}
+  doi = {10.1175/1520-0450(2003)042<0716:SOSLML>2.0.CO;2}
 }
 
 @article{pielke_1992_Meteorl.Atmos.Phys,
