+++ paper/paper.bib
@@ -254,20 +254,19 @@
 }
 
 @Article{Xue2000,
-author="Xue, M.
-and Droegemeier, K. K.
-and Wong, V.",
-title="The Advanced Regional Prediction System (ARPS) -- A multi-scale nonhydrostatic atmospheric simulation and prediction model. Part I: Model dynamics and verification",
-journal="Meteorology and Atmospheric Physics",
-year="2000",
-month="Dec",
-day="01",
-volume="75",
-number="3",
-pages="161--193",
-issn="1436-5065",
-doi="10.1007/s007030070003",
-url="https://doi.org/10.1007/s007030070003"
+  author="Xue, M.
+  and Droegemeier, K. K.
+  and Wong, V.",
+  title="The Advanced Regional Prediction System (ARPS) -- A multi-scale nonhydrostatic atmospheric simulation and prediction model. Part I: Model dynamics and verification",
+  journal="Meteorology and Atmospheric Physics",
+  year="2000",
+  month="Dec",
+  day="01",
+  volume="75",
+  number="3",
+  pages="161--193",
+  issn="1436-5065",
+  doi="10.1007/s007030070003"
 }
 
 
@@ -380,12 +379,12 @@ url="https://doi.org/10.1007/s007030070003"
 }
 
 @article{masson_2002_J.Appl.Meteorol,
-author = {Masson, V. and Grimmond, C. S. B. and Oke, T. R.},
-title = {Evaluation of the Town Energy Balance (TEB) Scheme with Direct Measurements from Dry Districts in Two Cities},
-journal = {Journal of Applied Meteorology},
-volume = {41},
-number = {10},
-pages = {1011-1026},
-year = {2002},
-doi = {10.1175/1520-0450(2002)041<1011:EOTTEB>2.0.CO;2}
+  author = {Masson, V. and Grimmond, C. S. B. and Oke, T. R.},
+  title = {Evaluation of the Town Energy Balance (TEB) Scheme with Direct Measurements from Dry Districts in Two Cities},
+  journal = {Journal of Applied Meteorology},
+  volume = {41},
+  number = {10},
+  pages = {1011-1026},
+  year = {2002},
+  doi = {10.1175/1520-0450(2002)041<1011:EOTTEB>2.0.CO;2}
 }
