+++ external/psychrolib/paper/paper.bib
@@ -1,0 +1,71 @@
+@book{Humphreys1920,
+  title={Physics of the Air},
+  author={Humphreys, W. J.},
+  year={1920},
+  publisher={Pub. for the Franklin Institute of the state of Pennsylvania by J.B. Lippincott Company},
+  address = {Philadelphia, PA}
+}
+
+@article{Dilley1968,
+  doi = {10.1175/1520-0450(1968)007<0717:otccov>2.0.co;2},
+  year  = {1968},
+  month = {aug},
+  publisher = {American Meteorological Society},
+  volume = {7},
+  number = {4},
+  pages = {717--719},
+  author = {Dilley, A. C.},
+  title = {On the Computer Calculation of Vapor Pressure and Specific Humidity Gradients from Psychrometric Data},
+  journal = {Journal of Applied Meteorology}
+}
+
+@book{Stoecker1982,
+  title={Refrigeration and Air Conditioning},
+  author={Stoecker, W.F. and Jones, J.W.},
+  isbn={9780070665910},
+  series={McGraw-Hill International Editions},
+  year={1982},
+  publisher={McGraw-Hill},
+  address = {New York, NY}
+}
+
+@book{Wexler1983,
+  title={Thermodynamic Properties of Dry Air, Moist Air and Water and SI Psychrometric Charts},
+  author={Wexler, A. and Hyland, R.W. and Stewart, R. and {American Society of Heating, Refrigerating and Air-Conditioning Engineers}},
+  year={1983},
+  publisher={ASHRAE},
+  address = {Atlanta, GA}
+}
+
+@article{Stull2011,
+  doi = {10.1175/jamc-d-11-0143.1},
+  year  = {2011},
+  month = {nov},
+  publisher = {American Meteorological Society},
+  volume = {50},
+  number = {11},
+  pages = {2267--2269},
+  author = {Roland Stull},
+  title = {Wet-Bulb Temperature from Relative Humidity and Air Temperature},
+  journal = {Journal of Applied Meteorology and Climatology}
+}
+
+@book{Ashrae2017_IP,
+  title={ASHRAE Handbook Fundamentals 2017: Inch-Pound System Edition},
+  author={{American Society of Heating, Refrigerating and Air-Conditioning Engineers}},
+  year={2017},
+  series={ASHRAE Handbook Fundamentals Inch-Pound System},
+  publisher={ASHRAE},
+  address = {Atlanta, GA},
+  isbn={9781939200570}
+}
+
+@book{Ashrae2017_SI,
+  title={ASHRAE Handbook Fundamentals 2017: International System Edition},
+  author={{American Society of Heating, Refrigerating and Air-Conditioning Engineers}},
+  year={2017},
+  series={ASHRAE Handbook Fundamentals International System},
+  publisher={ASHRAE},
+  address = {Atlanta, GA},
+  isbn={9781939200587}
+}
