+++ external/EnergyPlus/src/EnergyPlus/eplus_psychro.f90
@@ -1,0 +1,248 @@
+! MinimalDX version 0.1.1 (https://www.github.com/dmey/minimal-dx).
+! Copyright 2018 D. Meyer and R. Raustad. Licensed under MIT.
+
+! Description
+! This module is meant as a wrapper for the `PsyTsatFnHPb` function as it is missing from PsychroLib.
+! The `PsyTsatFnHPb` function was copied and pasted from the following link:
+! https://github.com/NREL/EnergyPlusRelease/blob/1ba84749/SourceCode/PsychRoutines.f90#L2331-L2542
+!
+! Changes
+! The following changes were made to the`§PsyTsatFnHPb` function:
+! - D. Meyer (20190112): replace EnergyPlus function `PsyHFnTdbW` and
+!   `PsyWFnTdbTwbPb` with `GetSatAirEnthalpy` from PsychroLib.
+! - D. Meyer (20190112): remove unused preprocessing directives.
+!
+! Copyright notice
+! EnergyPlus, Copyright (c) 1996-2019, The Board of Trustees of the University of Illinois,
+! The Regents of the University of California, through Lawrence Berkeley National Laboratory
+! (subject to receipt of any required approvals from the U.S. Dept. of Energy), Oak Ridge
+! National Laboratory, managed by UT-Battelle, Alliance for Sustainable Energy, LLC, and other
+! contributors. All rights reserved.
+!
+! NOTICE: This Software was developed under funding from the U.S. Department of Energy and the
+! U.S. Government consequently retains certain rights. As such, the U.S. Government has been
+! granted for itself and others acting on its behalf a paid-up, nonexclusive, irrevocable,
+! worldwide license in the Software to reproduce, distribute copies to the public, prepare
+! derivative works, and perform publicly and display publicly, and to permit others to do so.
+!
+! Redistribution and use in source and binary forms, with or without modification, are permitted
+! provided that the following conditions are met:
+!
+! (1) Redistributions of source code must retain the above copyright notice, this list of
+!     conditions and the following disclaimer.
+!
+! (2) Redistributions in binary form must reproduce the above copyright notice, this list of
+!     conditions and the following disclaimer in the documentation and/or other materials
+!     provided with the distribution.
+!
+! (3) Neither the name of the University of California, Lawrence Berkeley National Laboratory,
+!     the University of Illinois, U.S. Dept. of Energy nor the names of its contributors may be
+!     used to endorse or promote products derived from this software without specific prior
+!     written permission.
+!
+! (4) Use of EnergyPlus(TM) Name. If Licensee (i) distributes the software in stand-alone form
+!     without changes from the version obtained under this License, or (ii) Licensee makes a
+!     reference solely to the software portion of its product, Licensee must refer to the
+!     software as "EnergyPlus version X" software, where "X" is the version number Licensee
[…]

[…]
