+++ paper/paper.bib
@@ -388,3 +388,12 @@
   year = {2002},
   doi = {10.1175/1520-0450(2002)041<1011:EOTTEB>2.0.CO;2}
 }
+
+@Misc{kitware_inc._cmake_2019,
+  author   = {{Kitware Inc.}},
+  title    = {{CMake}},
+  year     = {2020},
+  language = {en-US},
+  url      = {https://cmake.org/},
+  urldate  = {2020-01-20},
+}
