+++ paper/paper.bib
@@ -375,7 +375,7 @@
   title = {{{WRF}}-{{TEB}}: Implementation and Evaluation of the Coupled {{Weather Research}} and {{Forecasting}} ({{WRF}}) and {{Town Energy Balance}} ({{TEB}}) Model},
   journal = {Submitted to Journal of Advances in Modeling Earth Systems},
   author = {Meyer, D. and Schoetter, R. and Riechert, M. and Verrelle, A. and Tewari, M. and Dudhia, J. and Masson, V. and {van Reeuwijk}, M. and Grimmond, S.},
-  year = {n.a.}
+  year = {2020}
 }
 
 @article{masson_2002_J.Appl.Meteorol,
