+++ CHANGELOG.txt
@@ -1,3 +1,6 @@
+0.1.2
+- Remove use of pure function in module MinimalDXFan to support older compilers.
+
 0.1.1
 - Psychrolib: Switch from git submodule to subtree.
 - Psychrolib: Update to version 2.4.0.
