### 2020-04-04 #2008 @dmey
Many thanks all for your comments and useful feedback. Below I respond to each of your points individually.

Reviewer: @ethan-nelson

> ## General checks
> * License: The OSI list only indicates that CECILL-2.1 is open source compatible. It looks like there are some differences between CECILL-1, and I'm not sure if the OSI designation is backwards-compatible, so I will defer to @mjsottile for that.

1.1: This has been changed to [CECILL-2.1](https://github.com/teb-model/teb/blob/master/Licence_CeCILL_V2.1-en.txt).

>
> ## Functionality
> * Installation: (CentOS 7.7, cmake 3.12.2, gcc 4.8.5) I ran into issues during compilation:
>
> ```
> /home/enelson/review/teb/external/minimal-dx/src/shared/minimal_dx_fan.f90:99.64:
>
>       error stop 'Fan Mode must either be 1 for on or 0 for off'
>                                                                 1
> Error: ERROR STOP statement not allowed in PURE procedure at (1)
> gmake[2]: *** [CMakeFiles/teb.dir/external/minimal-dx/src/shared/minimal_dx_fan.f90.o] Error 1
> gmake[1]: *** [CMakeFiles/teb.dir/all] Error 2
> gmake: *** [all] Error 2
> ```
>
> After removing the error message, the model compiled just fine. This is a fairly old gcc version though so it may not warrant a change.

1.2: This issue has been addressed directly in the external library. Can you please let me know if TEB now compiles OK on your system or open an issue?

> ## Documentation
> * Automated tests: this is a corner case, but it is useful if the user could specify a custom cmake/gcc/git binary location for tests, either in the python call or near the top of one of the python scripts. My cmake is not in my PATH variable so I had to manually change the location in the python code. Adding this would allow for faster iteration over different versions of gcc, etc. for testing compatibility.

1.3: The current test scripts are meant to be used for running regression tests and not as as general framework for running regular TEB cases. With regards to your specific problem, did you try adding CMake to your current shell before running CMake if you do not want to add it permanently to your PATH?

> * Community guidelines: Might I suggest a CONTRIBUTING.md file or indication in the README to complement the journal text? That could help direct folks who stumble upon the repo first as opposed to the paper.

1.4: `CONTRIBUTING.md` has been added.

> ## Other feedback:
> * A file in `docs/` with brief documentation detailing the necessary `input/` forcing files would be useful for tinkering from scratch. I ended up copying files from the example to play around with, but it would be good to have a list all the file variables and units--similar to the namelist docs.

1.5: This information has been added. Please see [input](https://github.com/teb-model/teb/blob/master/docs/software-docs.md#input).

> * Something else to consider: what are your thoughts on writing out a simple metadata file in the `output/` directory of a given run? Perhaps this onus should fall on the user to document their runs, but I could see potentially pulling in some metadata from the model build time (e.g. `git rev-parse`) that you could copy to `output/META.txt` or the like.

1.6: A TEB case directory comprises of a namelist, and input and output files. To retrieve metadata from the namelist we currently use the `get_date_params` function (see `teb/examples/CAPITOUL/tutorial.ipynb`). I am not yet sure what other metadata we should be saving but we could start discussing this in an issue for a future release.

Reviewer: @jayten

> I have gone through the review process for the [`teb`](https://github.com/teb-model/teb) repository. I was able to install (on Ubuntu18 and MacOs 10.13) and run test cases with instructions provided in the documentation. The software solves the intended problem well. The software is well written and documented. However, the flavor of the FORTRAN code style is old, owing to the use of old legacy code. For example, "impnone.fh" just contain an 'implicit none' statement. Although code style is not a major issue, it can be improved over time with the contribution of others. I would suggest @dmey to add, at least a simple, **CONTRIBUTING.md** for interested individuals to contribute to [`teb`](https://github.com/teb-model/teb).

2.1: Please see response to @ethan-nelson #1.4.

>
> # License
> I have issues with the license file: [CeCILL-C-V1](https://github.com/teb-model/teb/blob/joss-paper/Licence_CeCILL-C_V1-en.txt). The license file mentioned is not a part of the [OSI](https://opensource.org/licenses/alphabetical). Additionally, some old source files (like '[modei_flxsurf3db.f](https://github.com/teb-model/teb/blob/joss-paper/src/teb/modi_flxsurf3bx.f)') mention different license in the header, which is also not part of the [OSI](https://opensource.org/licenses/alphabetical). I want @mjsottile to take a decision on this issue.

2.2: Please see response to @ethan-nelson #1.1.

> # Paper
> @dmey , make sure all the contributors are at least acknowledged in the paper.md. For example, **P. Le Moigne**, mentioned as the author of the [_abort1_sfx.f90_](https://github.com/teb-model/teb/blob/joss-paper/src/driver/abor1_sfx.F90) source file, is neither mentioned in the author list nor the acknowledgment.

2.3: I have double-checked this with the other authors and the list of TEB contributors under the acknowledgments is reported correctly.

> ## References
> Missing DOI for bib entry: rozoff_2003_J.Appl.Meteorol

2.4: This has been added.

Editor/AEiC: @mjsottile/@danielskatz

All points should have been addressed. I only need to update one entry in the reference list (for which I am awaiting the DOI). With regards to Zenodo, I have added the full list of TEB contributors in the `zenodo.json` file. With regards to the licence, please see my response to @ethan-nelson #1.1.
