+++ README.md
@@ -56,7 +56,7 @@ This section includes links to software and model documentation. If you are new 
 
 ### Software
 
-See the [software documentation](docs/software-docs.md) for a general overview. For configuration options avalable at runtime please refer to [namelist options](docs/namelist-options.md).
+See the [software documentation](docs/software-docs.md) for a general overview. For configuration options available at runtime please refer to [namelist options](docs/namelist-options.md).
 
 ### Scientific
 The complete scientific documentation is included with the [SURFEX scientific documentation](http://www.umr-cnrm.fr/surfex/IMG/pdf/surfex_scidoc_v8.1.pdf).
@@ -73,7 +73,7 @@ The API documentation and code browser is available at [https://teb-model.github
 
 ## Testing
 
-Tests are run using the CAPITOUL data provided in `examples/CAPITOUL`. Tests are found in `tests/tests.py` and an overview is given in [`tests/README.md`](tests/README.md). All tests are automatically run at every commit using Continous Integration. If you are looking to run your tests locally on Linux or macOS, first make sure you have installed all the [prerequisites](#prerequisites), then from the command prompt:
+Tests are run using the CAPITOUL data provided in `examples/CAPITOUL`. Tests are found in `tests/tests.py` and an overview is given in [`tests/README.md`](tests/README.md). All tests are automatically run at every commit using Continuos Integration. If you are looking to run your tests locally on Linux or macOS, first make sure you have installed all the [prerequisites](#prerequisites), then from the command prompt:
 
 ``` bash
 python -m pip install -r requirements.txt
@@ -106,7 +106,7 @@ The corresponding reference list should be as follows:
 > XXXX *In preparation*
 
 
-*please make sure to cite the same version you are using with the correct DOI. For a list of all avalable versions see the list of versions on [Zenodo](XXXX).
+*please make sure to cite the same version you are using with the correct DOI. For a list of all available versions see the list of versions on [Zenodo](XXXX).
 
 
 ## Contributing
