### 2020-04-05 #2008 @mjsottile
Excellent!  It looks like everything has been addressed, but I do want to check on two items:

1. Can @ethan-nelson verify that 1.2 is addressed, and if not, open an issue as suggested by @dmey?

2. The license looks like it is addressed, but there is the caveat that @jayten pointed out that some of the files are from a third-party and have their own license wording in the comments (the "Environment Canada - Atmospheric Science and Technology License/Disclaimer").  The wording looks pretty permissive, but I am not a licensing expert.  Does @danielskatz have any input on this given that he's managed many more JOSS submissions than I have?

Thanks for your hard work on this @dmey.
