### 2020-04-13 #2008 @danielskatz
> the caveat that @jayten pointed out that some of the files are from a third-party and have their own license wording in the comments (the "Environment Canada - Atmospheric Science and Technology License/Disclaimer"). The wording looks pretty permissive, but I am not a licensing expert.

I'm afraid the answer is is that all software that is in the repo must have OSI-approved licenses.
