### 2020-04-14 #2008 @mjsottile
@dmey I'm afraid the inclusion of files from a third-party that are under a different license than your project is a problem.  Looking at the list of files that appear to have the Environment Canada license text (with corresponding line counts):

```
matt:teb/ (master) $ wc -l `grep -lnr Canada src`
      23 src/teb/modi_init_surfconsphy.f
      71 src/teb/init_surfconsphy.F
      73 src/teb/stabfunc2.fh
      32 src/teb/vslog.f
      17 src/teb/impnone.cdk
      18 src/teb/surfcon.fh
      28 src/teb/consphy.cdk
      73 src/teb/stabfunc2.cdk
      33 src/teb/modi_flxsurf3bx.f
      18 src/teb/surfcon.cdk
      28 src/teb/consphy.fh
      17 src/teb/impnone.fh
     251 src/teb/flxsurf3bx.F
```

That is a relatively tiny proportion of the overall code.  Do you have any thoughts on how to proceed?
