### 2020-04-14 #2008 @danielskatz
Just to be clear, the inclusion of files with a different license is not by itself the problem.

The problem is that
1. the licenses for all files needs to be OSI-approved (listed in https://opensource.org/licenses/alphabetical) - the Environment Canada license does not appear to be
2. licenses need to be compatible, which is a more complicated discussion
