### 2020-04-16 87e9c7a
Chnage licence to MIT for all MC2 files
