+++ requirements.txt
@@ -1,4 +1,4 @@
 numpy
 pandas
-f90nml
+f90nml==1.1.2
 matplotlib
