### 2020-04-18 fc945ff
Change licence to MIT for all MC2 files (#24)
Approved by Environment Canada. See email on 2020-04-17 at 20:15 UTC.
