+++ ThirdPartyNotices.txt
@@ -9,6 +9,7 @@ The original copyright notices and the licenses are included below:
 1.  dmey/minimal-dx version 0.1.1 (https://www.github.com/dmey/minimal-dx)
 2.  psychrometrics/psychrolib version 2.4.0 (https://github.com/psychrometrics/psychrolib)
 3.  NREL/EnergyPlusRelease version 8.1.0.009 (https://github.com/NREL/EnergyPlusRelease)
+4.  MC2 (Mesoscale Compressible Community) model
 
 %% dmey/minimal-dx NOTICES AND INFORMATION BEGIN HERE
 =========================================
@@ -112,3 +113,29 @@ OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED O
 POSSIBILITY OF SUCH DAMAGE.
 =========================================
 END OF NREL/EnergyPlusRelease NOTICES AND INFORMATION
+
+%% MC2 (Mesoscale Compressible Community) model NOTICES AND INFORMATION BEGIN HERE
+=========================================
+The MIT License (MIT)
+
+Copyright (c) 2001-2020 Environment Canada EC-RPN COMM Group.
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
+=========================================
+END OF MC2 (Mesoscale Compressible Community) model NOTICES AND INFORMATION
