### 2020-04-18 62fdfb0
remove minimaldx dep from repo and use CMake ExternalProject instead
