### 2020-04-18 c002dfd
Depend on MinimalDX through CMake's ExternalProject (#25)
Remove the redistribution of MinimalDX by relying on CMake's ExternalProject to download and build MinimalDX during TEB's build time.
