+++ external/minimal-dx/external/psychrolib/tests/test_psychrolib_si.py
@@ -1,220 +1,0 @@
-# PsychroLib (version 2.4.0) (https://github.com/psychrometrics/psychrolib).
-# Copyright (c) 2018-2020 The PsychroLib Contributors. Licensed under the MIT License.
-
-# Test of PsychroLib in SI units for Python, C, and Fortran.
-
-import numpy as np
-import pytest
-
-pytestmark = pytest.mark.usefixtures('SetUnitSystem_SI')
-
-# Test of helper functions
-def test_GetTKelvinFromTCelsius(psy):
-    assert psy.GetTKelvinFromTCelsius(20) == pytest.approx(293.15, 0.000001)
-
-def test_GetTCelsiusFromTKelvin(psy):
-    assert psy.GetTCelsiusFromTKelvin(293.15) == pytest.approx(20, rel = 0.000001)
-
-
-###############################################################################
-# Tests at saturation
-###############################################################################
-
-# Test saturation vapour pressure calculation
-# The values are tested against the values published in Table 3 of ch. 1 of the 2017 ASHRAE Handbook - Fundamentals
-# over the range [-100, +200] C
-# ASHRAE's assertion is that the formula is within 300 ppm of the true values, which is true except for the value at -60 C
-def test_GetSatVapPres(psy):
-    assert psy.GetSatVapPres(-60) == pytest.approx(1.08, abs = 0.01)
-    assert psy.GetSatVapPres(-20) == pytest.approx(103.24, rel = 0.0003)
-    assert psy.GetSatVapPres( -5) == pytest.approx(401.74, rel = 0.0003)
-    assert psy.GetSatVapPres(  5) == pytest.approx(872.6, rel = 0.0003)
-    assert psy.GetSatVapPres( 25) == pytest.approx(3169.7, rel = 0.0003)
-    assert psy.GetSatVapPres( 50) == pytest.approx(12351.3, rel = 0.0003)
-    assert psy.GetSatVapPres(100) == pytest.approx(101418.0, rel = 0.0003)
-    assert psy.GetSatVapPres(150) == pytest.approx(476101.4, rel = 0.0003)
-
-# Test saturation humidity ratio
-# The values are tested against those published in Table 2 of ch. 1 of the 2017 ASHRAE Handbook - Fundamentals
-# Agreement is not terrific - up to 2% difference with the values published in the table
-def test_GetSatHumRatio(psy):
-    assert psy.GetSatHumRatio(-50, 101325) == pytest.approx(0.0000243, rel = 0.01)
-    assert psy.GetSatHumRatio(-20, 101325) == pytest.approx(0.0006373, rel = 0.01)
-    assert psy.GetSatHumRatio( -5, 101325) == pytest.approx(0.0024863, rel = 0.005)
-    assert psy.GetSatHumRatio(  5, 101325) == pytest.approx(0.005425, rel = 0.005)
-    assert psy.GetSatHumRatio( 25, 101325) == pytest.approx(0.020173, rel = 0.005)
-    assert psy.GetSatHumRatio( 50, 101325) == pytest.approx(0.086863, rel = 0.01)
-    assert psy.GetSatHumRatio( 85, 101325) == pytest.approx(0.838105, rel = 0.02)
-
-# Test enthalpy at saturation
-# The values are tested against those published in Table 2 of ch. 1 of the 2017 ASHRAE Handbook - Fundamentals
-# Agreement is rarely better than 1%, and close to 3% at -5 C
-def test_GetSatAirEnthalpy(psy):
-    assert psy.GetSatAirEnthalpy(-50, 101325) == pytest.approx( -50222, rel = 0.01)
[…]
