+++ external/minimal-dx/external/psychrolib/src/js/psychrolib.js
@@ -1,1061 +1,0 @@
-/*
- * PsychroLib (version 2.4.0) (https://github.com/psychrometrics/psychrolib).
- * Copyright (c) 2018-2020 The PsychroLib Contributors for the current library implementation.
- * Copyright (c) 2017 ASHRAE Handbook — Fundamentals for ASHRAE equations and coefficients.
- * Licensed under the MIT License.
- */
-
-function Psychrometrics() {
-  /**
-   * Module overview
-   *  Contains functions for calculating thermodynamic properties of gas-vapor mixtures
-   *  and standard atmosphere suitable for most engineering, physical and meteorological
-   *  applications.
-   *
-   *  Most of the functions are an implementation of the formulae found in the
-   *  2017 ASHRAE Handbook - Fundamentals, in both International System (SI),
-   *  and Imperial (IP) units. Please refer to the information included in
-   *  each function for their respective reference.
-   *
-   * Example (e.g. Node.JS)
-   *  // Import the PsychroLib
-   *  var psychrolib = require('psychrolib.js')
-   *  // Set unit system
-   *  psychrolib.SetUnitSystem(psychrolib.SI)
-   *  // Calculate the dew point temperature for a dry bulb temperature of 25 C and a relative humidity of 80%
-   *  var TDewPoint = psychrolib.GetTDewPointFromRelHum(25.0, 0.80);
-   *  console.log('TDewPoint: %d', TDewPoint);
-   * 21.3094
-   *
-   * Copyright
-   *  - For the current library implementation
-   *     Copyright (c) 2018-2020 The PsychroLib Contributors.
-   *  - For equations and coefficients published ASHRAE Handbook — Fundamentals, Chapter 1
-   *     Copyright (c) 2017 ASHRAE Handbook — Fundamentals (https://www.ashrae.org)
-   *
-   * License
-   *  MIT (https://github.com/psychrometrics/psychrolib/LICENSE.txt)
-   *
-   * Note from the Authors
-   *  We have made every effort to ensure that the code is adequate, however, we make no
-   *  representation with respect to its accuracy. Use at your own risk. Should you notice
-   *  an error, or if you have a suggestion, please notify us through GitHub at
-   *  https://github.com/psychrometrics/psychrolib/issues.
-   */
-
-
-  // Standard functions
-  var log = Math.log;
-  var exp = Math.exp;
-  var pow = Math.pow;
-  var min = Math.min;
-  var max = Math.max;
-  var abs = Math.abs;
-
-
-  /******************************************************************************************************
-   * Global constants
-   *****************************************************************************************************/
-
-  var ZERO_FAHRENHEIT_AS_RANKINE = 459.67;  // Zero degree Fahrenheit (°F) expressed as degree Rankine (°R).
-                                            // Reference: ASHRAE Handbook - Fundamentals (2017) ch. 39.
-
-  var ZERO_CELSIUS_AS_KELVIN = 273.15;      // Zero degree Celsius (°C) expressed as Kelvin (K).
-                                            // Reference: ASHRAE Handbook - Fundamentals (2017) ch. 39.
-
-  var R_DA_IP = 53.350;               // Universal gas constant for dry air (IP version) in ft lb_Force lb_DryAir⁻¹ R⁻¹.
-                                      // Reference: ASHRAE Handbook - Fundamentals (2017) ch. 1.
-
[…]
