+++ external/minimal-dx/tests/debugging/README.md
@@ -1,2 +1,0 @@
-Contains a very simple program used as entry point to MinimalDX and EnergyPlus functions.
-Used for debugging purposes only.

[…]
