+++ tools/common_names.py
@@ -3,11 +3,11 @@
 
 standard_quantity_names = {
     "Atmospheric pressure": {
-        "unit": "hPa",
+        "unit": "\\rm Pa",
         "symbol": "p"
     },
     "Dry-bulb air temperature": {
-        "unit": "K",
+        "unit": "\\rm K",
         "symbol": "T"
     },
     "Relative humidity": {
@@ -15,107 +15,107 @@ standard_quantity_names = {
         "symbol": "\\rm RH"
     },
     "Total sky shortwave radiation flux density": {
-        "unit": "W\ m^{-2}",
+        "unit": "\\rm W\ m^{-2}",
         "symbol": "S⇊"
     },
     "Sky diffuse shortwave radiation flux density": {
-        "unit": "W\ m^{-2}",
+        "unit": "\\rm W\ m^{-2}",
         "symbol": "S⇓"
     },
     "Sky direct shortwave radiation flux density": {
-        "unit": "W\ m^{-2}",
+        "unit": "\\rm W\ m^{-2}",
         "symbol": "S↓"
     },
     "Surface reflected shortwave radiation flux density": {
-        "unit": "W\ m^{-2}",
+        "unit": "\\rm W\ m^{-2}",
         "symbol": "S↑"
     },
     "Sky longwave radiation flux density": {
-        "unit": "W\ m^{-2}",
+        "unit": "\\rm W\ m^{-2}",
         "symbol": "L↓"
     },
     "Surface longwave radiation flux density": {
-        "unit": "W\ m^{-2}",
+        "unit": "\\rm W\ m^{-2}",
         "symbol": "L↑"
     },
     "Surface net radiative flux density": {
-        "unit": "W\ m^{-2}",
+        "unit": "\\rm W\ m^{-2}",
         "symbol": "Q^*"
     },
     "Wind speed": {
-        "unit": "m\ s^{-1}",
+        "unit": "\\rm m\ s^{-1}",
         "symbol": "U"
     },
     "Wind direction": {
-        "unit": "rad",
+        "unit": "\\rm rad",
         "symbol": "U_{\\theta}"
     },
     "CO2 concentration": {
-        "unit": "g\ m^{-3}",
+        "unit": "\\rm g\ m^{-3}",
         "symbol": "CO_{2}"
     },
     "Mass mixing ratio of water vapour": {
-        "unit": "g\ kg^{-1}",
+        "unit": "\\rm g\ kg^{-1}",
         "symbol": "r"
     },
     "Surface turbulent sensible heat flux density": {
-        "unit": "W\ m^{-2}",
+        "unit": "\\rm W\ m^{-2}",
         "symbol": "H"
     },
     "Surface turbulent latent heat flux density": {
-        "unit": "W\ m^{-2}",
+        "unit": "\\rm W\ m^{-2}",
         "symbol": "LE"
     },
     "Rain rate": {
-        "unit": "mm\ h^{-1}",
+        "unit": "\\rm mm\ h^{-1}",
         "symbol": "RR"
     },
     "Buildings\' energy demand for cooling": {
-        "unit": "W\ m^{-2}",
+        "unit": "\\rm W\ m^{-2}",
         "symbol": "E_{\\rm cooling}"
     },
     "Buildings\' energy demand for heating": {
-        "unit": "W\ m^{-2}",
+        "unit": "\\rm W\ m^{-2}",
         "symbol": "E_{\\rm heating}"
     },
     "Thermal energy production of solar panels on roofs": {
-        "unit": "W\ m^{-2}",
+        "unit": "\\rm W\ m^{-2}",
         "symbol": "Q_{\\rm solar}"
     },
[…]
