### 2020-04-18 #2008 @dmey
> The license looks like it is addressed, but there is the caveat that @jayten pointed out that some of the files are from a third-party and have their own license wording in the comments (the "Environment Canada - Atmospheric Science and Technology License/Disclaimer"). The wording looks pretty permissive, but I am not a licensing expert. Does @danielskatz have any input on this given that he's managed many more JOSS submissions than I have?

@mjsottile @danielskatz @arfon. Environment Canada has kindly agreed to a change of licence for the files in question to MIT. I have now updated the repository to reflect this change.  Please could you let me know if you would like to see additional changes. Many thanks.
