### 2020-04-20 #2008 @mjsottile
@dmey That looks like good progress.  The final thing you should do is verify on the CeCILL license documentation that you have the appropriate license for TEB.  Specifically, the documentation surrounding "borrowings" (e.g., the Environment Canada code) here:

https://cecill.info/placer.en.html

It appears that CeCILL-B is compatible with MIT.  It is not clear that the regular CeCILL 2.1 is.  Please see https://en.wikipedia.org/wiki/CeCILL#Other_CeCILL_licenses for a bit more info.  Once the we've verified that your project has the correct CeCILL licensing to be compatible with the borrowed MIT-licensed code, we can move to the next steps of finalizing the review.
