### 2020-04-21 #2008 @dmey
@mjsottile Sorry but I am not sure what you mean exactly -- the Environment Canada (EC) code is licensed under MIT and bundled into TEB for redistribution (i.e. in a CeCILL 2.1 project). MIT is a (very) permissive licence with no copyleft.

In [How to distribute my program under CeCILL page](https://cecill.info/placer.en.html) under [Inventory of the pieces of code and libraries that you use](https://cecill.info/placer.en.html#emprunt), the second paragraph reads:

> To be able to distribute your software under a given license, it is necessary that the licenses of these borrowings be compatible with the chosen license i.e. they should not conflict with license. In the case of CeCILL, the licenses that are compatible with the GNU GPL (and that can be found on [this page](https://www.gnu.org/licenses/license-list.html)) are compatible with CeCILL.

From the quote above, from [this page](https://www.gnu.org/licenses/license-list.html), the [Expat License](https://www.gnu.org/licenses/license-list.html#Expat) (i.e. MIT) is shown under the list of "GPL-Compatible Free Software Licenses".
