### 2020-04-21 #2008 @mjsottile
@dmey Ok - I just wanted to make sure we had addressed @danielskatz second point re: license compatibility.  To my eye, it looks like we've addressed the license issue.  @danielskatz do you agree?  Assuming we're good I'll start my final pass over the review and materials.
