### 2020-04-21 #2008 @mjsottile
@dmey Looks good.  The only issue I see with the paper are some capitalizations in the references - the Rozoff title should have St. Louis, Missouri capitalized (it appears "st. Louis, missouri" in the PDF).  Similarly, the Xue paper title has bad capitalization.  I think in both cases it's just a matter of strategic placement of curly braces in the BibTeX title fields to keep the case correct in the generated output.

- Submit a version of the software to an archive (e.g., figshare or zenodo) and post the DOI for the archive in this thread.

- Please verify that the version number matches that of the archived code, and if not, let me know so I can update the version here.

- Finally, make sure that the metadata for the archive (e.g., title, authors) matches the paper.

Once all of these steps are done, I'll pass the paper on up the line to the EIC team for finalizing the process.
