### 2020-04-21 #2008 @danielskatz
Re licenses, can @dmey please explain (maybe one more time, sorry), which licenses are currently used in the repository?
