+++ CMakeLists.txt
@@ -56,7 +56,6 @@ set(src_driver
     src/driver/ol_time_interp_atm.F90
     src/driver/open_close_bin_asc_forc.F90
     src/driver/read_surf_atm.F90
-
     src/driver/modd_wrf_teb_driver.F90
 )
 
