### 2020-04-22 a123e25
remove static building onyl
