+++ external/minimal-dx/CMakeLists.txt
@@ -5,7 +5,7 @@ include(ExternalProject)
 
 ExternalProject_Add(minimaldx-project
     GIT_REPOSITORY https://github.com/dmey/minimal-dx.git
-    GIT_TAG 0.1.3
+    GIT_TAG dmey/cmake
     CMAKE_ARGS
         -DCMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE}
         -DCMAKE_Fortran_COMPILER=${CMAKE_Fortran_COMPILER}
