### 2020-04-22 8c91ac5
CMake: allow to build libteb as static/shared
