+++ paper/paper.bib
@@ -230,7 +230,7 @@
 
 @article{rozoff_2003_J.Appl.Meteorol,
   author = {Rozoff, Christopher M. and Cotton, William R. and Adegoke, Jimmy O.},
-  title = {Simulation of St. Louis, Missouri, Land Use Impacts on Thunderstorms},
+  title = {Simulation of {{St. Louis, Missouri}}, Land Use Impacts on Thunderstorms},
   journal = {Journal of Applied Meteorology},
   volume = {42},
   number = {6},
@@ -254,21 +254,20 @@
 }
 
 @Article{Xue2000,
-  author="Xue, M.
+  author = {Xue, M.
   and Droegemeier, K. K.
-  and Wong, V.",
-  title="The Advanced Regional Prediction System (ARPS) -- A multi-scale nonhydrostatic atmospheric simulation and prediction model. Part I: Model dynamics and verification",
-  journal="Meteorology and Atmospheric Physics",
-  year="2000",
-  month="Dec",
-  day="01",
-  volume="75",
-  number="3",
-  pages="161--193",
-  issn="1436-5065",
-  doi="10.1007/s007030070003"
+  and Wong, V.},
+  title={The {{Advanced Regional Prediction System (ARPS}}) -- A multi-scale nonhydrostatic atmospheric simulation and prediction model. {{Part I}}: Model dynamics and verification},
+  journal = {Meteorology and Atmospheric Physics},
+  year = {2000},
+  month = {Dec},
+  day = {01},
+  volume = {75},
+  number = {3},
+  pages = {161--193},
+  issn = {1436-5065},
+  doi = {10.1007/s007030070003}
 }
-
 
 @article{gerard_2009_Mon.Wea.Rev,
   title = {Cloud and {{Precipitation Parameterization}} in a {{Meso}}-{{Gamma}}-{{Scale Operational Weather Prediction Model}}},
