+++ src/teb/mode_psychrolib.f90
@@ -1,0 +1,1504 @@
+! PsychroLib (version 2.5.0) (https://github.com/psychrometrics/psychrolib).
+! Copyright (c) 2018-2020 The PsychroLib Contributors for the current library implementation.
+! Copyright (c) 2017 ASHRAE Handbook — Fundamentals for ASHRAE equations and coefficients.
+! Licensed under the MIT License.
+
+module mode_psychrolib
+  !+ Module overview
+  !+  Contains functions for calculating thermodynamic properties of gas-vapor mixtures
+  !+  and standard atmosphere suitable for most engineering, physical, and meteorological
+  !+  applications.
+  !+
+  !+  Most of the functions are an implementation of the formulae found in the
+  !+  2017 ASHRAE Handbook - Fundamentals, in both International System (SI),
+  !+  and Imperial (IP) units. Please refer to the information included in
+  !+  each function for their respective reference.
+  !+
+  !+ Example
+  !+  use psychrolib, only: GetTDewPointFromRelHum, SetUnitSystem, SI
+  !+  ! Set the unit system, for example to SI (can be either 'SI' or 'IP')
+  !+  call SetUnitSystem(SI)
+  !+  ! Calculate the dew point temperature for a dry bulb temperature of 25 C and a relative humidity of 80%
+  !+  print *, GetTDewPointFromRelHum(25.0, 0.80)
+  !+ 21.3094
+  !+
+  !+ Copyright
+  !+  - For the current library implementation
+  !+     Copyright (c) 2018-2020 The PsychroLib Contributors.
+  !+  - For equations and coefficients published ASHRAE Handbook — Fundamentals, Chapter 1
+  !+     Copyright (c) 2017 ASHRAE Handbook — Fundamentals (https://www.ashrae.org)
+  !+
+  !+ License
+  !+  MIT (https://github.com/psychrometrics/psychrolib/LICENSE.txt)
+  !+
+  !+ Note from the Authors
+  !+  We have made every effort to ensure that the code is adequate, however, we make no
+  !+  representation with respect to its accuracy. Use at your own risk. Should you notice
+  !+  an error, or if you have a suggestion, please notify us through GitHub at
+  !+  https://github.com/psychrometrics/psychrolib/issues.
+  !+
+  !+ Modifications from original PsychroLib version 2.5.0
+  !+ 20200421: `psychrolib` -> `mode_psychrolib` to avoid clashes with psychrolib used in MinimalDX.
+  !+            PsychroLib is used by both TEB and MinimalDX but each use their own.            
+
+
+  implicit none
+
+  private
+  public :: IP
+  public :: SI
+  public :: SetUnitSystem
+  public :: GetUnitSystem
+  public :: isIP
+  public :: GetTRankineFromTFahrenheit
+  public :: GetTFahrenheitFromTRankine
+  public :: GetTKelvinFromTCelsius
+  public :: GetTCelsiusFromTKelvin
+  public :: GetTWetBulbFromTDewPoint
+  public :: GetTWetBulbFromRelHum
+  public :: GetRelHumFromTDewPoint
+  public :: GetRelHumFromTWetBulb
[…]
