+++ ThirdPartyNotices.txt
@@ -7,6 +7,7 @@ This project incorporates components from the projects listed below.
 The original copyright notices and the licenses are included below:
 
 1.  MC2 (Mesoscale Compressible Community) model
+2.  psychrometrics/psychrolib version 2.5.0 (https://github.com/psychrometrics/psychrolib)
 
 %% MC2 (Mesoscale Compressible Community) model NOTICES AND INFORMATION BEGIN HERE
 =========================================
@@ -33,3 +34,29 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 =========================================
 END OF MC2 (Mesoscale Compressible Community) model NOTICES AND INFORMATION
+
+%% psychrometrics/psychrolib NOTICES AND INFORMATION BEGIN HERE
+=========================================
+The MIT License (MIT)
+
+Copyright (c) 2018-2020 The PsychroLib Contributors.
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
+=========================================
+END OF psychrometrics/psychrolib NOTICES AND INFORMATION
