+++ examples/CAPITOUL/README.md
@@ -1,10 +1,24 @@
 # CAPITOUL case
 
-In this folder you will find all the data to run TEB (Masson, 2000) with data from the CAPITOUL (Canopy and Aerosol Particles Interactions in TOulouse Urban Layer; Masson et al., 2008) campaign.
+In this folder data are provided to run TEB (Masson, 2000) with sample input data from the CAPITOUL (Canopy and Aerosol Particles Interactions in TOulouse Urban Layer; Masson et al., 2008) campaign. Data are provided for demonstration purposes only and are not quality controlled (i.e. should not be used to conduct scientific experiments). Both data and tutorial are only meant to guide the user through setting up a simple experiment in TEB. If you are looking to conduct a scientific experiment, please download and process the data yourself (see the [software documentation page](../../docs/software-docs.md)) and amend the input namelist accordingly (see the [namelist options page](../../docs/namelist-options)). 
 
-## Example applications
+## Install Python packages
 
-Please refer to the [`tutorial.ipynb`](tutorial.ipynb) provided.
+Several Python packages are required to run the tutorial. To install these, execute the following from the command line:
+
+```
+pip install -r requirements.txt
+```
+
+## Tutorial
+
+To run the tutorial, execute the following from the command line:
+
+```
+jupyter notebook tutorial.ipynb
+```
+
+Alteratively, your can view its content in [`tutorial.ipynb`](tutorial.ipynb).
 
 
 ## References
