### 2020-04-23 #2008 @dmey
> @dmey Looks good. The only issue I see with the paper are some capitalizations in the references - the Rozoff title should have St. Louis, Missouri capitalized (it appears "st. Louis, missouri" in the PDF). Similarly, the Xue paper title has bad capitalization. I think in both cases it's just a matter of strategic placement of curly braces in the BibTeX title fields to keep the case correct in the generated output.

@mjsottile These have been fixed -- thank you. I am having an issue with the placement of Table 1 when this is typeset using the JOSS template. Would you know how to move Table 1 to page 2 so that the table is kept whole? I tried with a new line (using the HTML `<br>` tag) but that did not seem to have helped...

> Re licenses, can @dmey please explain (maybe one more time, sorry), which licenses are currently used in the repository?

Sure, no problem. To recap, the TEB project repository is licensed under CeCILL 2.1. The two 'external' codes that are currently bundled  (and redistributed) in TEB are: (a) subset of subroutines from the Mesoscale Compressible Community model from Environmental Canada (licensed under MIT) and (b) PsychroLib (licensed under MIT). The MinimalDX code is not bundled/redistributed with TEB as it is built and linked using CMake ExternalProject (as it usually done with other libraries components).

The two licences for the bundled code are in the `ThirdPartyNotices.txt` file.
