### 2020-04-23 #2008 @mjsottile
@dmey I didn't see any specific instructions in the JOSS docs but it does appear that you can use LaTeX commands to control formatting in pandoc (which is what whedon uses to generate the PDF).  See: https://stackoverflow.com/questions/16965490/pandoc-markdown-page-break

@arfon can weigh in - I think he knows Whedon best and would know how to do that.  Alternatively, you could break the paragraphs above the table up to force some whitespace to be inserted between paragraphs to force the table onto the next page and avoid inserting LaTeX commands into the markdown.

Once you're happy with the format of the paper, please see the points from my comment above (https://github.com/openjournals/joss-reviews/issues/2008#issuecomment-617407446) regarding archiving and we can finish the process.
