### 2020-04-23 #2008 @arfon
> @arfon can weigh in - I think he knows Whedon best and would know how to do that. Alternatively, you could break the paragraphs above the table up to force some whitespace to be inserted between paragraphs to force the table onto the next page and avoid inserting LaTeX commands into the markdown.

Yeah, I think you can issue LaTeX commands such as `\newpage` to control the document flow.
