### 2020-04-24 db7773d
Try forcing page 1 to end before table 1 begins
