+++ paper/paper.md
@@ -40,7 +40,7 @@ The Town Energy Balance (TEB) model [@masson_2000_Boundary-LayerMeteorology] is 
 
 Here, we present an enhanced software and platform for the TEB model to help scientists and practitioners wishing to use the TEB model in their research as a standalone software application or as a library in their own software. This includes several features such as cross-platform support for Windows, Linux, and macOS using CMake [@kitware_inc._cmake_2019], static and dynamic library generation for integration with other software/models, namelist-based configuration, integration with MinimalDX [@meyer_d_2019_3562311] and PsychroLib [@meyer_2019_JOSS] to improve the modelling of air conditioners (AC) and psychrometric calculations respectively, a thin interface used in the coupling with WRF-CMake [@riechert_2019_JOSS], helper functions for Python for pre- and post-processing inputs and outputs files, and a tutorial in Jupyter Notebook to allow users to quickly become familiar with the general TEB modeling workflow. In the new platform we implement testing at every code commit through continuous integration (CI) and automate the generation of API documentation. The project is developed as a free, open source, community-driven project on GitHub ([https://github.com/teb-model/teb](https://github.com/teb-model/teb)) to support existing and new model applications with enhanced functionality. We welcome contributions and encourage users to provide feedback, bug reports and feature requests, via GitHub's issue system at [https://github.com/teb-model/teb/issues](https://github.com/teb-model/teb/issues).
 
-<br />
+\newpage
 
 | Modeling capability                                                   | Reference                                                                 |
 | --------------------------------------------------------------------- | ------------------------------------------------------------------------- |
