### 2020-04-24 #2008 @dmey
Forcing page 1 to end with `\newpage` has worked and Table 1 is now rendered whole -- thanks.

> Once you're happy with the format of the paper, please see the points from my comment above ([#2008 (comment)](https://github.com/openjournals/joss-reviews/issues/2008#issuecomment-617407446)) regarding archiving and we can finish the process.

@mjsottile With regards to your last point in [#2008 (comment)](https://github.com/openjournals/joss-reviews/issues/2008#issuecomment-617407446), I added a [Zenodo metadata file](https://github.com/TEB-model/teb/blob/joss-paper/zenodo.json) to the repository. Can you please check that this is fine -- as we previously discussed, for Zenodo we wanted to keep consistent with previous and future version to avoid forking.

I am also waiting to hear back about a reference and DOI to be finalized so that I can update the list of references -- hopefully this will be done in the next couple of weeks. In the meantime, please let me know if there are other changes you would like to see.
