### 2020-05-12 #2008 @mjsottile
:wave: @dmey - Just checking to see if you've heard back with the information you were waiting on to see if we're ready to move forward.  I believe that the metadata issue is fine with respect to keeping the repository sane and avoid unnecessary forking.
