### 2020-06-09 2e91d3b
minor updates
