+++ paper/paper.bib
@@ -361,20 +361,19 @@
 @software{meyer_d_2019_3562311,
   author       = {Meyer, D. and Raustad, R.},
   title        = {MinimalDX},
-  month        = dec,
   year         = 2019,
   publisher    = {Zenodo},
-  version      = {0.1.0},
   doi          = {10.5281/zenodo.3562310},
   url          = {https://doi.org/10.5281/zenodo.3562310}
 }
 
 
-@article{meyer_n.a._Submitt.J.Adv.Model.EarthSyst,
+@article{meyer_2020_.J.Adv.Model.EarthSyst,
   title = {{{WRF}}-{{TEB}}: Implementation and Evaluation of the Coupled {{Weather Research}} and {{Forecasting}} ({{WRF}}) and {{Town Energy Balance}} ({{TEB}}) Model},
-  journal = {Submitted to Journal of Advances in Modeling Earth Systems},
+  journal = {Journal of Advances in Modeling Earth Systems},
   author = {Meyer, D. and Schoetter, R. and Riechert, M. and Verrelle, A. and Tewari, M. and Dudhia, J. and Masson, V. and {van Reeuwijk}, M. and Grimmond, S.},
-  year = {2020}
+  year = {2020},
+  doi = {10.1029/2019MS001961}
 }
 
 @article{masson_2002_J.Appl.Meteorol,
