+++ README.md
@@ -5,7 +5,7 @@
 
 [![Build Status](https://travis-ci.com/teb-model/teb.svg?branch=master)](https://travis-ci.com/teb-model/teb)
 
-[Overview](#overview) | [Prerequisites](#prerequisites) | [Installation](#installation) | [Documentation](#documentation) | [Example application](#example-application) | [Testing](#testing) | [How to cite](#how-to-cite) | [Contributing](#contributing) | [Copyright and license](#copyright-and-license) | [References](#references)
+[Overview](#overview) | [Prerequisites](#prerequisites) | [Installation](#installation) | [Documentation](#documentation) | [Example application](#example-application) | [Testing](#testing) | [How to cite](#how-to-cite) | [Contributing](#contributing) | [Copyright and license](#copyright-and-license)
 </div>
 
 ## Overview
@@ -25,8 +25,9 @@ This enhanced software and platform for TEB (Town Energy Balance; [Masson, 2000]
 
 To build the Town Energy Balance (TEB) executable and library on Windows, Linux and macOS, clone this repository and run the following commands from your command prompt:
 
-```bash
-mkdir build && cd build
+```
+mkdir build
+cd build
 cmake -DCMAKE_BUILD_TYPE=Release ..
 cmake --build .
 ```
@@ -68,7 +69,7 @@ The code browser is available at [https://teb-model.github.io/teb](https://teb-m
 
 ## Example application
 
-[`tutorial.ipynb`](examples/CAPITOUL/tutorial.ipynb) provides a simple tutorial detailing how TEB can be used to to estimate the buildings' energy demand for cooling using sample data and parameters from the CAPITOUL (Canopy and Aerosol Particles Interactions in TOulouse Urban Layer; Masson et al., 2008) campaign.
+To get started with TEB, see [`examples/CAPITOUL/README.md`](examples/CAPITOUL/README.md). The folder contains a simple tutorial in [Jupyter Notebook](https://jupyter.org/) to estimate the buildings' energy demand for cooling using TEB and data and parameters from the CAPITOUL (Canopy and Aerosol Particles Interactions in TOulouse Urban Layer; [Masson et al., 2008](https://doi.org/10.1007/s00703-008-0289-4)) campaign.
 
 
 ## Testing
@@ -94,19 +95,18 @@ python tests/test.py --build_type=Debug --case=integration
 
 When using the TEB software, please cite both model, and software (with version) as follows:
 
-| Model        | Software | Version*                   |
-| ------------ | -------- | -------------------------- |
-| Masson, 2000 | XXXX     | [see Zenodo](https://XXXX) |
+| Physical model                                          | This software                                      | Version*       |
+| ------------------------------------------------------- | -------------------------------------------------- | -------------- |
+| [Masson, 2000](https://doi.org/10.1023/A:1002463829265) | [Meyer, 2020](https://doi.org/10.21105.joss.02008) | [see Zenodo]() |
 
 
 The corresponding reference list should be as follows:
 
 > Masson, V., 2000: A Physically-Based Scheme For The Urban Energy Budget In Atmospheric Models. Boundary-Layer Meteorology, 94, 357–397, https://doi.org/10.1023/A:1002463829265.
 
-> XXXX *In preparation*
+> Meyer, D., Schoetter, R., Masson, V., Grimmond, S., 2020: Enhanced software and platform for the Town Energy Balance (TEB) model. Journal of Open Source Software. https://doi.org/10.21105.joss.02008.
 
-
-*please make sure to cite the same version you are using with the correct DOI. For a list of all available versions see the list of versions on [Zenodo](XXXX).
[…]
