+++ CONTRIBUTING.md
@@ -1,5 +1,5 @@
 # How to contribute
 
-Thank you for considering contributing. In general, you can contribute by reporting an issue or by directly contributing to the source code. For the latter, fork our repository, clone the fork, make your changes, and create a pull request (PR) with a **clear description** of your changes -- if you are unfamiliar about forking/creating PR, please see [this guide](https://guides.github.com/activities/forking/) first. If/when your changes are merged, you will appear as one of our [Contributors](https://github.com/teb-model/teb/graphs/contributors). If you would like to know more about specific development guidelines and testing, please refer to our [README.md](README.md) file.
+Thank you for considering contributing. In general, you can contribute by reporting an issue or by directly contributing to the source code. For the latter, fork our repository, clone the fork, make your changes, and create a pull request (PR) with a **clear description** of your changes -- if you are unfamiliar about forking/creating PRs, please see [this guide](https://guides.github.com/activities/forking/) first. If/when your changes are merged, you will appear as one of our [Contributors](https://github.com/teb-model/teb/graphs/contributors). If you would like to know more about specific development guidelines and testing, please refer to our [README.md](README.md) file.
 
 If you would like to report a bug, please check if similar issue have already been reported at [https://github.com/teb-model/teb/issues](https://github.com/teb-model/teb/issues). If none exist please create a new issue, including as many details as possible about your system and issue encountered.
