### 2020-06-09 #2008 @mjsottile
Excellent.  Once you've posted the archive DOI and checked that the version number of the archive matches the version listed at the top of this thread, I'll run the final commands to hand off to the EIC team for the last steps.
