+++ paper/paper.bib
@@ -1,0 +1,397 @@
+@article{Redon2020,
+  doi = {10.5194/gmd-13-385-2020},
+  url = {https://doi.org/10.5194/gmd-13-385-2020},
+  year = {2020},
+  month = feb,
+  publisher = {Copernicus {GmbH}},
+  volume = {13},
+  number = {2},
+  pages = {385--399},
+  author = {Emilie Redon and Aude Lemonsu and Val{\'{e}}ry Masson},
+  title = {An urban trees parameterization for modeling microclimatic variables and thermal comfort conditions at street level with the Town Energy Balance model ({TEB}-{SURFEX} v8.0)},
+  journal = {Geoscientific Model Development}
+}
+
+@article{Redon2017,
+  doi = {10.5194/gmd-10-385-2017},
+  url = {https://doi.org/10.5194/gmd-10-385-2017},
+  year = {2017},
+  month = jan,
+  publisher = {Copernicus {GmbH}},
+  volume = {10},
+  number = {1},
+  pages = {385--411},
+  author = {Emilie C. Redon and Aude Lemonsu and Val{\'{e}}ry Masson and Benjamin Morille and Marjorie Musy},
+  title = {Implementation of street trees within the solar radiative exchange parameterization of {TEB} in {SURFEX} v8.0},
+  journal = {Geoscientific Model Development}
+}
+
+@article{Pigeon2014,
+  doi = {10.1016/j.enbuild.2013.10.038},
+  url = {https://doi.org/10.1016/j.enbuild.2013.10.038},
+  year = {2014},
+  month = jun,
+  publisher = {Elsevier {BV}},
+  volume = {76},
+  pages = {1--14},
+  author = {G. Pigeon and K. Zibouche and B. Bueno and J. Le Bras and V. Masson},
+  title = {Improving the capabilities of the Town Energy Balance model with up-to-date building energy simulation algorithms: an application to a set of representative buildings in Paris},
+  journal = {Energy and Buildings}
+}
+
+@article{masson_2000_Boundary-LayerMeteorology,
+  title = {A {{Physically}}-{{Based Scheme For The Urban Energy Budget In Atmospheric Models}}},
+  volume = {94},
+  issn = {0006-8314, 1573-1472},
+  language = {en},
+  number = {3},
+  journal = {Boundary-Layer Meteorology},
+  doi = {10.1023/A:1002463829265},
+  author = {Masson, Val{\'e}ry},
+  month = mar,
+  year = {2000},
+  pages = {357-397}
+}
+
+@article{lemonsu_2004_J.Appl.Meteorol,
+  title = {Modeling the {{Surface Energy Balance}} of the {{Core}} of an {{Old Mediterranean City}}: {{Marseille}}},
+  volume = {43},
+  language = {en},
+  journal = {Journal of Applied Meteorology},
+  doi = {10.1175/1520-0450(2004)043<0312:MTSEBO>2.0.CO;2},
+  author = {Lemonsu, A and Grimmond, C S B and Masson, V},
+  year = {2004},
+  pages = {16}
+}
+
+@article{lemonsu_2012_Geosci.ModelDev,
+  title = {Inclusion of Vegetation in the {{Town Energy Balance}} Model for Modelling Urban Green Areas},
+  volume = {5},
+  issn = {1991-9603},
+  language = {en},
+  number = {6},
+  journal = {Geoscientific Model Development},
+  doi = {10.5194/gmd-5-1377-2012},
[…]
