+++ paper/paper.md
@@ -1,0 +1,65 @@
+---
+title: 'Enhanced software and platform for the Town Energy Balance (TEB) model'
+tags:
+  - Meteorology
+  - Fortran
+  - Energy
+  - Urban
+  - Energy Balance
+  - Python
+  - HVAC
+
+authors:
+  - name: D. Meyer
+    orcid: 0000-0002-7071-7547
+    affiliation: "1, 2"
+  - name: R. Schoetter
+    orcid: 0000-0002-2284-4592
+    affiliation: 3
+  - name: V. Masson
+    orcid: 0000-0001-8807-0545
+    affiliation: 3
+  - name: S. Grimmond
+    orcid: 0000-0002-3166-9415
+    affiliation: 1
+affiliations:
+ - name: Department of Meteorology, University of Reading, Reading, UK
+   index: 1
+ - name: Department of Civil and Environmental Engineering, Imperial College London, London, UK
+   index: 2
+ - name: CNRM UMR 3589, Université Fédérale de Toulouse, Météo-France/CNRS, Toulouse, France
+   index: 3
+date: 04 December 2019
+bibliography: paper.bib
+---
+
+# Summary
+
+The Town Energy Balance (TEB) model [@masson_2000_Boundary-LayerMeteorology] is a physically based single layer Urban Canopy Model (UCM) to calculate the urban surface energy balance at neighborhood scale assuming a simplified canyon geometry. It includes several capabilities (Table 1) that have been extensively evaluated offline with flux observations [@leroyer_2010_J.Appl.Meteor.Climatol; @pigeon_2008_MeteorolAtmosPhys; @lemonsu_2004_J.Appl.Meteorol; @masson_2002_J.Appl.Meteorol] and online coupled to atmospheric models such as ALARO [@gerard_2009_Mon.Wea.Rev] in ALARO-TEB [@hamdi_2012_Wea.Forecasting], the Global Environmental Multiscale (GEM; Côté et al. [-@cote_1998_Mon.WeatherRev]) in GEM-TEB [@lemonsu_2009_Boundary-LayerMeteorol], Meso-NH [@lafore_1998_Ann.Geophys; @lac_2018_Geosci.ModelDev] in TEB-MesoNH [@lemonsu_2002_Boundary-LayerMeteorology], the Regional Atmospheric Modeling System (RAMS; Pielke et al. [-@pielke_1992_Meteorl.Atmos.Phys]) in RAMS-TEB [@freitas_2007_Boundary-LayerMeteorol], the Advanced Regional Prediction System (ARPS; Xue et al., [-@Xue2000]) in ARPS-TEB [@rozoff_2003_J.Appl.Meteorol], and the Weather Research and Forecasting  (WRF; Skamarock et al. [-@skamarock_2019_NCARTech.NoteNCARTN-556STR]) in WRF-TEB [@meyer_2020_.J.Adv.Model.EarthSyst].
+
+
[…]
