### 2020-06-09 #2008 @kyleniemeyer
Hi @dmey - can you confirm the DOI 10.1029/2019MS001961 for your 2020 JAMES article? Is that an upcoming article that is not online yet?
