### 2020-06-09 #2008 @dmey
Hi @kyleniemeyer. That is correct -- the DOI was sent out to me today but it may take a few more days before it becomes active.
